requirements.txt
streamlit
torch
transformers
sentence-transformers
faiss-cpu
PyPDF2
pyttsx3
SpeechRecognition
pyaudio
numpy
tokenizers